+++ src/audio_output.rs
@@ -1,6 +1,5 @@
 use crate::audio::{
     Audio, AudioCommand, AudioCommandResult, InstanceHandle, PlayAudioSettings, PlaybackState,
-    PlaybackStatus,
 };
 use bevy::prelude::*;
 use bevy::utils::tracing::warn;
@@ -152,25 +151,17 @@ impl AudioOutput {
             .expect("Failed to add arrangement to the AudioManager")
     }
 
-    fn stop(
-        &mut self,
-        channel: &AudioChannel,
-        audio_states: &mut HashMap<InstanceHandle, PlaybackState>,
-    ) -> AudioCommandResult {
+    fn stop(&mut self, channel: &AudioChannel) -> AudioCommandResult {
         if let Some(instances) = self.instances.get_mut(channel) {
-            while !instances.is_empty() {
-                let mut instance = instances.remove(0);
+            for instance in instances {
                 match instance.kira.stop(StopInstanceSettings::default()) {
-                    Ok(()) => {
-                        audio_states.remove(&instance.handle);
-                    }
                     Err(CommandError::CommandQueueFull) => {
-                        instances.push(instance);
                         return AudioCommandResult::Retry;
                     }
                     Err(error) => {
                         println!("Failed to stop instance: {:?}", error);
                     }
+                    _ => (),
                 }
             }
         }
@@ -332,7 +323,7 @@ impl AudioOutput {
                         AudioCommandResult::Retry
                     }
                 }
-                AudioCommand::Stop => self.stop(&channel, &mut audio.states),
+                AudioCommand::Stop => self.stop(&channel),
                 AudioCommand::Pause => {
                     self.pause(&channel);
                     AudioCommandResult::Ok
@@ -466,19 +457,17 @@ pub fn update_instance_states_system(world: &mut World) {
     let mut audio = world.get_resource_mut::<Audio>().unwrap();
     for instance_state_vec in audio_output.instances.values() {
         for instance_state in instance_state_vec.iter() {
-            let playback_state = PlaybackState {
-                status: match instance_state.kira.state() {
-                    KiraInstanceState::Playing => PlaybackStatus::Playing,
-                    KiraInstanceState::Paused(_) => PlaybackStatus::Paused,
-                    KiraInstanceState::Stopped => PlaybackStatus::Stopped,
-                    KiraInstanceState::Pausing(_) => PlaybackStatus::Pausing,
-                    KiraInstanceState::Stopping => PlaybackStatus::Stopping,
-                },
-                position: Some(instance_state.kira.position()),
+            let position = instance_state.kira.position();
+            let playback_status = match instance_state.kira.state() {
+                KiraInstanceState::Playing => PlaybackState::Playing { position },
+                KiraInstanceState::Paused(_) => PlaybackState::Paused { position },
+                KiraInstanceState::Stopped => PlaybackState::Stopped,
+                KiraInstanceState::Pausing(_) => PlaybackState::Pausing { position },
+                KiraInstanceState::Stopping => PlaybackState::Stopping { position },
             };
             audio
                 .states
-                .insert(instance_state.handle.clone(), playback_state);
+                .insert(instance_state.handle.clone(), playback_status);
         }
     }
 }